added integration with Stripe for course payments and endpoint description in Swagger

+++ materials/models.py
@@ -7,7 +7,7 @@ class Course(models.Model):
     title = models.CharField(max_length=255, verbose_name="Название")
     preview = models.ImageField(upload_to="courses/", blank=True, null=True, verbose_name="Превью")
     description = models.TextField(verbose_name="Описание")
-    price = models.PositiveIntegerField(default=0)
+    price = models.DecimalField(max_digits=10, decimal_places=2, default=0)
 
     def __str__(self):
         return self.title
